refactor(tests): Moved the test files for manifest system to the correct directory
